+++ code/packages/rust/closure-pass-inline/src/lib.rs
@@ -580,6 +580,13 @@ fn count_decl_names_stmt(
                 }
                 count_decl_names_stmt(&fs.body, out, nodes_touched);
             }
+            TaggedStatement::ForOfStatement(fs) => {
+                // The for-in `left`, when a declaration, binds the loop variable.
+                if let ForInit::VariableDeclaration(vd) = &fs.left {
+                    count_decl_names_var(vd, out);
+                }
+                count_decl_names_stmt(&fs.body, out, nodes_touched);
+            }
             TaggedStatement::LabeledStatement(ls) => {
                 count_decl_names_stmt(&ls.body, out, nodes_touched)
             }
@@ -797,6 +804,20 @@ fn tally_stmt(stmt: &Statement, cand: &InlineCandidate, t: &mut Tally) {
                 tally_stmt(&fs.body, cand, t);
             }
             TaggedStatement::ForInStatement(fs) => {
+                match &fs.left {
+                    ForInit::VariableDeclaration(vd) => {
+                        for d in &vd.declarations {
+                            if let Some(i) = &d.init {
+                                tally_expr(i, cand, t);
+                            }
+                        }
+                    }
+                    ForInit::Expression(e) => tally_expr(e, cand, t),
+                }
+                tally_expr(&fs.right, cand, t);
+                tally_stmt(&fs.body, cand, t);
+            }
+            TaggedStatement::ForOfStatement(fs) => {
                 match &fs.left {
                     ForInit::VariableDeclaration(vd) => {
                         for d in &vd.declarations {
@@ -1037,6 +1058,20 @@ fn inline_in_stmt(stmt: &mut Statement, cand: &InlineCandidate) -> bool {
                 changed |= inline_in_stmt(&mut fs.body, cand);
             }
             TaggedStatement::ForInStatement(fs) => {
+                match &mut fs.left {
+                    ForInit::VariableDeclaration(vd) => {
+                        for d in &mut vd.declarations {
+                            if let Some(i) = &mut d.init {
+                                changed |= inline_in_expr(i, cand);
+                            }
+                        }
+                    }
+                    ForInit::Expression(e) => changed |= inline_in_expr(e, cand),
+                }
+                changed |= inline_in_expr(&mut fs.right, cand);
+                changed |= inline_in_stmt(&mut fs.body, cand);
+            }
+            TaggedStatement::ForOfStatement(fs) => {
                 match &mut fs.left {
                     ForInit::VariableDeclaration(vd) => {
                         for d in &mut vd.declarations {
@@ -1953,6 +1988,9 @@ fn splice_void_in_stmt(
             TaggedStatement::ForInStatement(fs) => {
                 splice_void_in_slot(&mut fs.body, cand, avoid, nodes_touched)
             }
+            TaggedStatement::ForOfStatement(fs) => {
+                splice_void_in_slot(&mut fs.body, cand, avoid, nodes_touched)
+            }
             TaggedStatement::LabeledStatement(ls) => {
                 splice_void_in_slot(&mut ls.body, cand, avoid, nodes_touched)
             }
@@ -2795,6 +2833,9 @@ fn splice_valued_in_stmt(
             TaggedStatement::ForInStatement(fs) => {
                 splice_valued_in_stmt(&mut fs.body, cand, avoid, nodes_touched)
             }
+            TaggedStatement::ForOfStatement(fs) => {
+                splice_valued_in_stmt(&mut fs.body, cand, avoid, nodes_touched)
+            }
             TaggedStatement::LabeledStatement(ls) => {
                 splice_valued_in_stmt(&mut ls.body, cand, avoid, nodes_touched)
             }
@@ -3106,6 +3147,20 @@ fn collect_used_idents_stmt(stmt: &Statement, out: &mut HashSet<String>) {
                 collect_used_idents_stmt(&fs.body, out);
             }
             TaggedStatement::ForInStatement(fs) => {
+                match &fs.left {
+                    ForInit::VariableDeclaration(vd) => {
+                        for d in &vd.declarations {
+                            if let Some(i) = &d.init {
+                                collect_binding_idents_expr(i, out);
+                            }
+                        }
+                    }
+                    ForInit::Expression(e) => collect_binding_idents_expr(e, out),
+                }
+                collect_binding_idents_expr(&fs.right, out);
+                collect_used_idents_stmt(&fs.body, out);
+            }
+            TaggedStatement::ForOfStatement(fs) => {
                 match &fs.left {
                     ForInit::VariableDeclaration(vd) => {
                         for d in &vd.declarations {